+++ bundle/terranova/tnresources/app.go
@@ -101,16 +101,3 @@ func (r *ResourceApp) waitForApp(ctx context.Context, w *databricks.WorkspaceCli
 		}
 	})
 }
-
-// This is copied from the retries package of the databricks-sdk-go. It should be made public,
-// but for now, I'm copying it here.
-func shouldRetry(err error) bool {
-	if err == nil {
-		return false
-	}
-	e := err.(*retries.Err)
-	if e == nil {
-		return false
-	}
-	return !e.Halt
-}
