+++ bundle/terranova/tnresources/util.go
@@ -2,6 +2,8 @@ package tnresources
 
 import (
 	"reflect"
+
+	"github.com/databricks/databricks-sdk-go/retries"
 )
 
 // filterFields creates a new slice with fields present only in the provided type.
@@ -17,4 +19,17 @@ func filterFields[T any](fields []string) []string {
 	}
 
 	return result
+}
+
+// This is copied from the retries package of the databricks-sdk-go. It should be made public,
+// but for now, I'm copying it here.
+func shouldRetry(err error) bool {
+	if err == nil {
+		return false
+	}
+	e := err.(*retries.Err)
+	if e == nil {
+		return false
+	}
+	return !e.Halt
 }
